Fix bug with overriding options object
Recently, with the addition of supporting running LSP as a standalone
process, the `connect` function was changed.

This changed the promise such that the options object was now the result
of the `connection.onInitialized` function rather than the result of
the callback from `connection.onInitialize`, which seems like an error
in understanding the API of the LSP library

This reverts that change, but still returns the value options out of the
promise, though it is likely an uneeded change.

Automated testing for this would be ideal, but as there doesn't seem to
be any infrastructure for that currently, that is out of scope of this
commit